# Inline the PostCSS config file into the webpack config