--- a/client/postcss.config.js
+++ b/client/postcss.config.js
@@ -1,9 +1,0 @@
-module.exports = ( { options } ) => {
-	const plugins = { autoprefixer: {} };
-	if ( options.transformCssProperties ) {
-		plugins[ 'postcss-custom-properties' ] = {
-			importFrom: [ options.customProperties ],
-		};
-	}
-	return { plugins };
-};