@scinos 2020-08-24
For completeness, I don't get the performance benefits in my system: the fallback build tooks about 2 minutes in all cases:


```
# Before

Commit: 594a8badc9 with `postbuild-client-fallback` disabled

## Cold caches

Command: rm -fr .cache/ .build/ public/fallback && time NODE_ENV=production CALYPSO_ENV=production yarn run build-client-fallback
Average: 1:56.29
Raw data:
	NODE_ENV=production CALYPSO_ENV=production yarn run build-client-fallback  154.81s user 14.33s system 143% cpu 1:58.07 total
	NODE_ENV=production CALYPSO_ENV=production yarn run build-client-fallback  151.49s user 13.56s system 144% cpu 1:54.51 total
	NODE_ENV=production CALYPSO_ENV=production yarn run build-client-fallback  151.26s user 13.93s system 142% cpu 1:55.66 total


## Warm caches

Command: time NODE_ENV=production CALYPSO_ENV=production yarn run build-client-fallback
Average: 2:04.99
Raw data:
	NODE_ENV=production CALYPSO_ENV=production yarn run build-client-fallback  154.96s user 15.02s system 142% cpu 1:59.66 total
	NODE_ENV=production CALYPSO_ENV=production yarn run build-client-fallback  173.44s user 18.74s system 143% cpu 2:13.74 total
	NODE_ENV=production CALYPSO_ENV=production yarn run build-client-fallback  153.43s user 14.80s system 138% cpu 2:01.57 total

---

# After

Commit: 7986ad73d0 with `postbuild-client-fallback` disabled

## Cold caches

Command: rm -fr .cache/ .build/ public/fallback && time NODE_ENV=production CALYPSO_ENV=production yarn run build-client-fallback
Average: 2:06.74
Raw data:
	NODE_ENV=production CALYPSO_ENV=production yarn run build-client-fallback  166.93s user 19.75s system 143% cpu 2:09.66 total
	NODE_ENV=production CALYPSO_ENV=production yarn run build-client-fallback  175.55s user 17.54s system 143% cpu 2:14.49 total
	NODE_ENV=production CALYPSO_ENV=production yarn run build-client-fallback  152.63s user 14.64s system 144% cpu 1:56.08 total

## Warm caches

Command: time NODE_ENV=production CALYPSO_ENV=production yarn run build-client-fallback
Average: 1:57.70
Raw data:
	NODE_ENV=production CALYPSO_ENV=production yarn run build-client-fallback  151.40s user 13.24s system 144% cpu 1:54.24 total
	NODE_ENV=production CALYPSO_ENV=production yarn run build-client-fallback  151.96s user 12.99s system 144% cpu 1:54.06 total
	NODE_ENV=production CALYPSO_ENV=production yarn run build-client-fallback  165.02s user 16.86s system 145% cpu 2:04.79 total
```