@jsnajdr 2020-08-24
@scinos I see that when testing with "cold caches", you use the following command to delete the caches: `rm -fr .cache/ .build/ public/fallback`

On my machine, I also have a `~/.cache` folder in my home directory, with a `webpack/css` subfolder. The results of the CSS compilations seem to be cached there, and they use the PostCSS pipeline that is being optimized in this PR. The subfolder is regenerated on the next build after I delete it, so it's not something stale. It seems that your "cold caches" build was using that cache, as it was not delete?

Is it expected that the `~/.cache` folder is used, and used only by the CSS compilations? It seems that something in the `mini-css-extract-plugin` configuration was omitted when you recently worked on the `cache-loader` and the `.cache` directory.