@scinos 2020-08-24
Data with a colder cache:

```
# Before

Commit: 594a8badc9 with `postbuild-client-fallback` disabled

## Colder caches

Command: rm -fr ~/.cache/webpack .cache/ .build/ public/fallback && time NODE_ENV=production CALYPSO_ENV=production yarn run build-client-fallback
Average: 3:06.13
Raw data:
	NODE_ENV=production CALYPSO_ENV=production yarn run build-client-fallback  252.58s user 22.27s system 140% cpu 3:15.05 total
	NODE_ENV=production CALYPSO_ENV=production yarn run build-client-fallback  232.14s user 18.25s system 142% cpu 2:55.69 total
	NODE_ENV=production CALYPSO_ENV=production yarn run build-client-fallback  238.04s user 22.88s system 139% cpu 3:07.30 total

---

# After

Commit: 7986ad73d0 with `postbuild-client-fallback` disabled

## Colder caches

Command: rm -fr ~/.cache/webpack .cache/ .build/ public/fallback && time NODE_ENV=production CALYPSO_ENV=production yarn run build-client-fallback
Average: 2:59.43
Raw data:
	NODE_ENV=production CALYPSO_ENV=production yarn run build-client-fallback  231.28s user 21.88s system 131% cpu 3:12.93 total
	NODE_ENV=production CALYPSO_ENV=production yarn run build-client-fallback  217.50s user 19.49s system 140% cpu 2:48.96 total
	NODE_ENV=production CALYPSO_ENV=production yarn run build-client-fallback  228.40s user 21.66s system 141% cpu 2:56.40 total
```

The diff is ~7 seconds. Not sure if it is statistically relevant, given that I got runs with a difference of >20 seconds.