@jsnajdr 2020-08-24
> The diff is ~7 seconds. Not sure if it is statistically relevant, given that I got runs with a difference of >20 seconds.

Thanks for testing! The 7 seconds diff is not that great, but still nice if it proves to be real.

The patch removes a code path that loads a small JS file from filesystem and executes it, repeated 800 times. 7 seconds looks like a plausible timing for that.